+++ src/data_inclusion/schema/models.py
@@ -7,11 +7,11 @@ from pydantic import BaseModel, EmailStr, Extra, HttpUrl, constr
 from data_inclusion.schema.code_officiel_geographique import TypeCOG
 from data_inclusion.schema.frais import Frais
 from data_inclusion.schema.labels_nationaux import LabelNational
-from data_inclusion.schema.modalites_orientation import (
-    ModaliteOrientationAccompagnateur,
-    ModaliteOrientationBeneficiaire,
-)
 from data_inclusion.schema.modes_accueil import ModeAccueil
+from data_inclusion.schema.modes_orientation import (
+    ModeOrientationAccompagnateur,
+    ModeOrientationBeneficiaire,
+)
 from data_inclusion.schema.profils import Profil
 from data_inclusion.schema.thematiques import Thematique
 from data_inclusion.schema.typologies_de_services import TypologieService
@@ -60,10 +60,8 @@ class Service(BaseModel):
     ]
     zone_diffusion_nom: Optional[str]
     contact_nom: Optional[str]
-    modalites_orientation_beneficiaire: Optional[list[ModaliteOrientationBeneficiaire]]
-    modalites_orientation_accompagnateur: Optional[
-        list[ModaliteOrientationAccompagnateur]
-    ]
+    modes_orientation_beneficiaire: Optional[list[ModeOrientationBeneficiaire]]
+    modes_orientation_accompagnateur: Optional[list[ModeOrientationAccompagnateur]]
 
     class Config:
         extra = Extra.forbid