+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 * schéma service:
   * champs ajoutés:
     * `contact_nom`
-    * `modalites_orientation_beneficiaire`
-    * `modalites_orientation_accompagnateur`
+    * `modes_orientation_beneficiaire`
+    * `modes_orientation_accompagnateur`
 * typologies de structures:
   * ajout crèches AVIP (`AVIP`)
 * thématiques: